tools/maskedmail/pyproject.toml
[project]
name = "maskedmail"
version = "0.1"
description = ""
readme = "README.md"
authors = [{name = "Logan Linn"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "jmapc>=0.2.22",
    "onepassword-sdk>=0.1.7",
    "requests>=2.32.3",
    "rich>=13.9.4",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
maskedmail= "maskedmail.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
